4.7 into master

+++ brand_safety/management/commands/import_new_tags.py
@@ -7,6 +7,7 @@ from brand_safety.models import BadWordCategory
 from audit_tool.models import AuditLanguage
 from brand_safety.languages import LANG_CODES
 from utils.utils import remove_tags_punctuation
+from utils.lang import is_english
 
 
 logger = logging.getLogger(__name__)
@@ -19,12 +20,33 @@ class Command(BaseCommand):
             help="Filename of newly vetted tags."
         )
 
+        parser.add_argument(
+            "--delete_all",
+            help="Set to True if you want to delete all tags in the database and reimport."
+        )
+
+        parser.add_argument(
+            "--overwrite",
+            help="Default value is False. Set to True if you want to overwrite duplicate tags when importing."
+        )
+
     def handle(self, *args, **kwargs):
         file_name = kwargs["file_name"]
         invalid_rows_file_name = "invalid_new_tags.csv"
         invalid_rows = []
+        try:
+            if kwargs["delete_all"]:
+                BadWord.objects.all().delete()
+        except Exception:
+            pass
+        try:
+            if kwargs["overwrite"]:
+                overwrite = True
+            else:
+                overwrite = False
+        except Exception:
+            overwrite = False
 
-        BadWord.objects.all().delete()
         counter = 0
         with open(file_name, "r") as f:
             reader = csv.reader(f)
@@ -35,8 +57,8 @@ class Command(BaseCommand):
                 try:
                     # Parse word
                     word = remove_tags_punctuation(row[0].lower().strip())
-                    if len(word) < 3:
-                        reason = [f"Word {word} is shorter than 3 characters long when trimmed."]
+                    if is_english(word) and len(word) < 3:
+                        reason = [f"Word {word} is shorter than 3 English characters long when trimmed."]
                         invalid_rows.append(row + reason)
                         continue
                     # Parse category
@@ -52,6 +74,17 @@ class Command(BaseCommand):
                         reason = ["One of word, category, language, or negative_score not found."]
                         invalid_rows.append(row + reason)
                         continue
+
+                    if not overwrite:
+                        try:
+                            bad_word = BadWord.objects.get(name=word, language=language)
+                            reason = [f"'overwrite' parameter is set to False, but the word '{word}' with language "
+                                      f"'{language}' already exists in the database."]
+                            invalid_rows.append(row + reason)
+                            continue
+                        except Exception as e:
+                            pass
+
                     try:
                         bad_word = BadWord.all_objects.get(name=word, language=language)
                         bad_word.deleted_at = None
